rename RelNodeUtil to RelExplainUtil and do some refactor
