+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/calcite/Rank.scala
@@ -172,7 +172,7 @@ case class VariableRankRange(rankEndIndex: Int) extends RankRange {
 object Rank {
   def sortFieldsToString(collationSort: RelCollation): String = {
     val fieldCollations = collationSort.getFieldCollations
-      .map(c => (c.getFieldIndex, RelNodeUtil.directionToOrder(c.getDirection)))
+      .map(c => (c.getFieldIndex, FlinkRelOptUtil.directionToOrder(c.getDirection)))
 
     fieldCollations.map {
       case (index, order) => s"$$$index ${order.getShortName}"
@@ -181,7 +181,7 @@ object Rank {
 
   def sortFieldsToString(collationSort: RelCollation, inputType: RelDataType): String = {
     val fieldCollations = collationSort.getFieldCollations
-      .map(c => (c.getFieldIndex, RelNodeUtil.directionToOrder(c.getDirection)))
+      .map(c => (c.getFieldIndex, FlinkRelOptUtil.directionToOrder(c.getDirection)))
     val inputFieldNames = inputType.getFieldNames
 
     fieldCollations.map {
