+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/common/CommonExchange.scala
@@ -16,11 +16,12 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.plan.nodes.physical.common
+package org.apache.flink.table.plan.nodes.common
 
 import org.apache.flink.table.plan.`trait`.FlinkRelDistribution
 import org.apache.flink.table.plan.cost.FlinkCost._
 import org.apache.flink.table.plan.cost.FlinkCostFactory
+import org.apache.flink.table.plan.nodes.FlinkRelNode
 
 import org.apache.calcite.plan.{RelOptCluster, RelOptCost, RelOptPlanner, RelTraitSet}
 import org.apache.calcite.rel.core.Exchange
@@ -30,14 +31,15 @@ import org.apache.calcite.rel.{RelDistribution, RelNode, RelWriter}
 import scala.collection.JavaConverters._
 
 /**
-  * Base class for physical [[Exchange]].
+  * Base class for flink [[Exchange]].
   */
 abstract class CommonExchange(
     cluster: RelOptCluster,
     traitSet: RelTraitSet,
     relNode: RelNode,
     relDistribution: RelDistribution)
-  extends Exchange(cluster, traitSet, relNode, relDistribution) {
+  extends Exchange(cluster, traitSet, relNode, relDistribution)
+  with FlinkRelNode {
 
   override def computeSelfCost(planner: RelOptPlanner, mq: RelMetadataQuery): RelOptCost = {
     val inputRows = mq.getRowCount(input)
