+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/logical/FlinkLogicalSort.scala
@@ -19,7 +19,7 @@
 package org.apache.flink.table.plan.nodes.logical
 
 import org.apache.flink.table.plan.nodes.FlinkConventions
-import org.apache.flink.table.plan.util.RelNodeUtil
+import org.apache.flink.table.plan.util.{FlinkRelOptUtil, RelExplainUtil}
 
 import org.apache.calcite.plan._
 import org.apache.calcite.rel.convert.ConverterRule
@@ -43,7 +43,7 @@ class FlinkLogicalSort(
   extends Sort(cluster, traits, child, collation, offset, fetch)
   with FlinkLogicalRel {
 
-  private lazy val limitStart: Long = RelNodeUtil.getLimitStart(offset)
+  private lazy val limitStart: Long = FlinkRelOptUtil.getLimitStart(offset)
 
   override def copy(
       traitSet: RelTraitSet,
