+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/physical/batch/BatchExecSortAggregate.scala
@@ -18,7 +18,7 @@
 package org.apache.flink.table.plan.nodes.physical.batch
 
 import org.apache.flink.table.functions.UserDefinedFunction
-import org.apache.flink.table.plan.util.RelNodeUtil
+import org.apache.flink.table.plan.util.RelExplainUtil
 
 import org.apache.calcite.plan.{RelOptCluster, RelTraitSet}
 import org.apache.calcite.rel._
@@ -75,10 +75,10 @@ class BatchExecSortAggregate(
     super.explainTerms(pw)
       .item("isMerge", isMerge)
       .itemIf("groupBy",
-        RelNodeUtil.fieldToString(grouping, inputRowType), grouping.nonEmpty)
+        RelExplainUtil.fieldToString(grouping, inputRowType), grouping.nonEmpty)
       .itemIf("auxGrouping",
-        RelNodeUtil.fieldToString(auxGrouping, inputRowType), auxGrouping.nonEmpty)
-      .item("select", RelNodeUtil.groupAggregationToString(
+        RelExplainUtil.fieldToString(auxGrouping, inputRowType), auxGrouping.nonEmpty)
+      .item("select", RelExplainUtil.groupAggregationToString(
         inputRowType,
         outputRowType,
         grouping,
